Add .astype() conversion to TableModel.__init__()

+++ gammapy/spectrum/models.py
@@ -81,7 +81,7 @@ class SpectralModel(object):
             Tuple of flux and flux error.
         """
         energy = self._convert_energy(energy)
-        
+
         unit = self(energy).unit
         upars = self.parameters._ufloats
         uarray = self.evaluate(energy.value, **upars)
@@ -132,7 +132,7 @@ class SpectralModel(object):
 
         def f(x):
             return self.evaluate(x, **upars)
-        
+
         uarray = integrate_spectrum(f, emin.value, emax.value, **kwargs)
         return self._parse_uarray(uarray) * unit
 
@@ -170,7 +170,7 @@ class SpectralModel(object):
             Lower bound of integration range.
         emax : `~astropy.units.Quantity`
             Upper bound of integration range
-        
+
         Returns
         -------
         energy_flux, energy_flux_error : tuple of `~astropy.units.Quantity`
@@ -179,13 +179,13 @@ class SpectralModel(object):
         """
         emin = self._convert_energy(emin)
         emax = self._convert_energy(emax)
-        
+
         unit = self.energy_flux(emin, emax, **kwargs).unit
         upars = self.parameters._ufloats
 
         def f(x):
             return x * self.evaluate(x, **upars)
-        
+
         uarray = integrate_spectrum(f, emin.value, emax.value, **kwargs)
         return self._parse_uarray(uarray) * unit
 
@@ -250,7 +250,7 @@ class SpectralModel(object):
         flux = self(energy).to(flux_unit)
 
         y = self._plot_scale_flux(energy, flux, energy_power)
-        
+
         ax.plot(energy.value, y.value, **kwargs)
         self._plot_format_ax(ax, energy, y, energy_power)
         return ax
@@ -288,7 +288,7 @@ class SpectralModel(object):
         kwargs.setdefault('facecolor', 'black')
         kwargs.setdefault('alpha', 0.2)
         kwargs.setdefault('linewidth', 0)
-        
+
         emin, emax = energy_range
         energy = EnergyBounds.equal_log_spacing(
             emin, emax, n_points, energy_unit)
@@ -297,7 +297,7 @@ class SpectralModel(object):
 
         y_lo = self._plot_scale_flux(energy, flux - flux_err, energy_power)
         y_hi = self._plot_scale_flux(energy, flux + flux_err, energy_power)
-        
+
         where = (y_hi > 0) & (energy >= energy_range[0]) & (energy <= energy_range[1])
         ax.fill_between(energy.value, y_lo.value, y_hi.value, where=where, **kwargs)
         self._plot_format_ax(ax, energy, y_lo, energy_power)
@@ -311,7 +311,7 @@ class SpectralModel(object):
             ax.set_ylabel('Flux [{}]'.format(y.unit))
         ax.set_xscale("log", nonposx='clip')
         ax.set_yscale("log", nonposy='clip')
-        
+
     def _plot_scale_flux(self, energy, flux, energy_power):
         eunit = [_ for _ in flux.unit.bases if _.physical_type == 'energy'][0]
 
@@ -785,8 +785,10 @@ class TableModel(SpectralModel):
                 y = np.log10(self.values)
             else:
                 y = self.values
-        self.interpy = interp1d(loge,
-                                y,
+        # The type conversion is a fix for:
+        # https://travis-ci.org/gammapy/gammapy/jobs/210576260
+        self.interpy = interp1d(loge.astype(float),
+                                y.astype(float),
                                 fill_value=-np.Inf,
                                 bounds_error=False,
                                 kind='cubic')
